[core] Wrap list of errors stored in an Errors in struct.
Instead of aliasing it to a plain array. This would force the error
checking to use `Errors.HaveOccurred()` instead of comparing against nil
-- which is incorrect because an Errors contains an empty array of
Error's should be considered "no errors".

Also add numerous methods to make use of Errors easier and intuitive.

Finally add a matcher `testutil.ConsistOfGraphQLErrors` to match
`Errors` (as it no longer an array but a struct).