[types] Remove ThisIsGraphQL*TypeDefinition marker interface.
`ThisIsTypeDefinition` is kept to make `TypeDefinition` a non-trivial
interface.

TypeDefinition usually contains an interface TypeData which returns a
value (e.g., ScalarTypeData) which is only meaningful to NewType and is
not used elsewhere. Removing it to make define TypeDefinition easier and
increase code coverage.

This requires renaiming `ElementType` to `InnerType` in
`NonNullTypeDefinition` to distinguish it from `ListTypeDefinition`.

Closes #86.