[executor] ExecuteNodeTask as a ResolveInfo.
`executor.ExecuteNodeTask` now implements `graphql.ResolveInfo` so we
can pass the task itself to field resolve when appropriated without
creating an `executor.ResolveInfo` object. This is a memory
optimization. When resolving value for task.result (that's the case for
`ExecuteNodeTask.run`), we can pass:

```go
info := &ResolveInfo{
        ExecutionContext: task.ctx,
        ExecutionNode:    task.node,
        ResultNode:       task.result,
}
```

But a better way is to use "task" directly as an `ResolveInfo` object to
save allocation overheads.